+++ bin/vinyld/cache/cache_vrt.c
@@ -888,43 +888,6 @@ VRT_Rollback(VRT_CTX, VCL_HTTP hp)
 
 /*--------------------------------------------------------------------*/
 
-VCL_VOID
-VRT_synth_strands(VRT_CTX, VCL_STRANDS s)
-{
-	int i, hasnull = 0;
-	const char * const s_null = "(null)";
-
-	CHECK_OBJ_NOTNULL(s, STRANDS_MAGIC);
-	for (i = 0; i < s->n; i++) {
-		if (s->p[i] != NULL)
-			continue;
-		s->p[i] = s_null;
-		hasnull = 1;
-	}
-
-	VRT_l_resp_body(ctx, LBODY_ADD_STRING, NULL, s);
-	if (hasnull == 0)
-		return;
-	for (i = 0; i < s->n; i++) {
-		if (s->p[i] == s_null)
-			s->p[i] = NULL;
-	}
-}
-
-VCL_VOID
-VRT_synth_blob(VRT_CTX, VCL_BLOB b)
-{
-	VRT_l_resp_body(ctx, LBODY_SET_BLOB, NULL, b);
-}
-
-VCL_VOID
-VRT_synth_page(VRT_CTX, VCL_STRANDS s)
-{
-	VRT_synth_strands(ctx, s);
-}
-
-/*--------------------------------------------------------------------*/
-
 static VCL_STRING
 vrt_ban_error(VRT_CTX, VCL_STRING err)
 {